Test q_shuffle fairly
Prvious,I using srand() only once in the beginning of test,which means
each shuffle depends on the previous rand() calls. That could make the
shuffle look more random than it really is.Now we set a different seed
each time we shuffle, to make sure the shuffle logic is tested fairly.

Change-Id: Iae319cec6f79a7ed1e1b9fcdce839070f435630b